refactor(rdf-store-rs): fixing lint issues

+++ src/c_utils.rs
@@ -4,9 +4,10 @@
 use std::ffi::CStr;
 use crate::RDFStoreError;
 
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
 pub fn ptr_to_cstr<'b>(data: *const u8, len: usize) -> Result<&'b CStr, RDFStoreError> {
     unsafe {
-        let slice = std::slice::from_raw_parts(data, len as usize);
+        let slice = std::slice::from_raw_parts(data, len);
         Ok(CStr::from_bytes_with_nul_unchecked(slice))
     }
 }
